requirements.txt
pydantic
streamlit
pandas
boto3
envarify
requests
python-dotenv